[hotfix][streaming] Fix and simplify PrintSinkFunctionTest
Previously PrintSinkFunctionTest was testing incorrectly for System.err. setTargetToStandardErr was being
called after opening PrintSink so it didn't have any effect.

This commit also reduce code duplication and get rids of mockito usage in this test.
